- fixed bug with invalid transmission of short or int (little to big endian) * now correctly sending captured mat from client to server

+++ RacecARGO/VMMRRequest.h
@@ -18,6 +18,6 @@
 
 @interface VMMRRequest : TCPRequest
 
-- (void)startWithDescriptors:(const uint8_t*)descriptors withRows:(int)rows andCols:(int)cols;
+- (void)startWithDescriptors:(const uint8_t*)descriptors withRows:(uint16_t)rows andCols:(uint16_t)cols;
 
 @end
